pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cleanepi"
version = "0.1.0"
authors = [
    {name = "Converted from R package by epiverse-trace team"},
]
description = "Clean and standardize epidemiological data"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "python-dateutil>=2.8.0",
    "pyjanitor>=0.20.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.910",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]

[project.urls]
Homepage = "https://github.com/asembereng/cleanepi_py"
Documentation = "https://github.com/asembereng/cleanepi_py"
Repository = "https://github.com/asembereng/cleanepi_py"
"Bug Tracker" = "https://github.com/asembereng/cleanepi_py/issues"

[tool.setuptools]
packages = ["cleanepi"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]